# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GymPal Backend API",
    "description": "Complete API for the GymPal fitness application with social features, personal tracking, workouts, and analytics",
    "version": "1.0.0",
    "contact": {
      "name": "GymPal Team",
      "email": "support@gympal.app",
      "url": "https://gympal.app"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://api.gympal.app",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/workouts/{id}/exercises": {
      "get": {
        "tags": ["Workouts"],
        "summary": "Get workout exercises",
        "description": "Retrieve all exercises in a workout",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workout exercises retrieved successfully"
          },
          "404": {
            "description": "Workout not found"
          }
        }
      },
      "post": {
        "tags": ["Workouts"],
        "summary": "Add exercise to workout",
        "description": "Add an exercise to a workout",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddExerciseRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Exercise added successfully"
          },
          "404": {
            "description": "Workout not found"
          }
        }
      }
    },
    "/api/v1/workouts/{id}/exercises/{exerciseId}": {
      "put": {
        "tags": ["Workouts"],
        "summary": "Update workout exercise",
        "description": "Update an exercise in a workout",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "exerciseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExerciseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Exercise updated successfully"
          },
          "404": {
            "description": "Exercise not found"
          }
        }
      },
      "delete": {
        "tags": ["Workouts"],
        "summary": "Remove exercise from workout",
        "description": "Remove an exercise from a workout",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "exerciseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Exercise removed successfully"
          },
          "404": {
            "description": "Exercise not found"
          }
        }
      }
    },
...
    "/api/v1/exercises": {
      "get": {
        "tags": ["Exercises"],
        "summary": "List exercises",
        "description": "Get paginated list of exercises with optional filtering",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "muscleGroup",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "equipment",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Exercises retrieved successfully"
          }
        }
      }
    },
    "/api/v1/exercises/{id}": {
      "get": {
        "tags": ["Exercises"],
        "summary": "Get exercise by ID",
        "description": "Retrieve detailed information about a specific exercise",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Exercise retrieved successfully"
          },
          "404": {
            "description": "Exercise not found"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "AddExerciseRequest": {
        "type": "object",
        "required": ["exerciseId"],
        "properties": {
          "exerciseId": {
            "type": "string",
            "format": "uuid"
          },
          "sets": {
            "type": "integer",
            "minimum": 1,
            "example": 3
          },
          "reps": {
            "type": "integer",
            "minimum": 1,
            "example": 10
          },
          "weight": {
            "type": "number",
            "minimum": 0,
            "example": 70.5
          },
          "order": {
            "type": "integer",
            "minimum": 0,
            "example": 1
          }
        }
      },
      "UpdateExerciseRequest": {
        "type": "object",
        "properties": {
          "sets": {
            "type": "integer",
            "minimum": 1
          },
          "reps": {
            "type": "integer",
            "minimum": 1
          },
          "weight": {
            "type": "number",
            "minimum": 0
          },
          "order": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
    }
  }
}